Update plate-temperature.py
printed every intermediate value to check where was the fault.
forgot why S was coming out to be incorrect, but fixed it afterall.
expression of RaL was wrong and messy also, so fixed that. That was causing the error in calculation.
Now the code is complete and I checked it with the example in book and getting expected answers.